requirements.txt
fastapi
openai
uvicorn[standard]
python-multipart
python-dotenv
anthropic
sentence-transformers
huggingface_hub>=0.27.0
httpx
ranx
pypdf
pdfplumber
psutil
pynvml
numpy
scikit-learn
